trim back commit to just include st.experimental_rerun()
